feat: Moved ApplicationFlag.java to a nested enum inside of Application.java, and added the AutoMod badge flag.
